refactor(album): Refactor album feature to its own app
Refactor albu feature to its own app, including model, views(detail, create, update). \nAnd refactor some media handling logic into mixin for AlbumForm and PostForm
