+++ core/network/albums/admin.py
@@ -1,0 +1,11 @@
+"""Album admin."""
+
+from django.contrib import admin
+
+from .models import Album
+
+
+# Register your models here.
+@admin.register(Album)
+class AlbumAdmin(admin.ModelAdmin):
+    """Custom Album Admin."""
