+++ core/network/urls.py
@@ -10,4 +10,5 @@ urlpatterns = [
     path("accounts/", include("allauth.urls")),
     path("profiles/", include("network.profiles.urls")),
     path("posts/", include("network.posts.urls")),
+    path("albums/", include("network.albums.urls")),
 ]
